+++ packages/client/functions/moon.js
@@ -1,5 +1,0 @@
-import moonObject from "../../framework/shared/moon-object";
-
-(function() {
-    window.moon = moonObject(window.moonConfig);
-})();