+++ start.js
@@ -1,6 +1,6 @@
 /**
  * Imports the startServer method from the build folders. Only works
- * after moon-js has been transpiled.
+ * after luna-js has been transpiled.
  */
 const { startServer } = require("./lib/packages/framework");
 startServer();